header init
