+++ packages/nextjs/components/Test.tsx
@@ -1,0 +1,19 @@
+import BiconomyButton from "./BiconomyButton";
+import FarmerCard from "./FarmerCard";
+import IPFSUploader from "./IPFSUploader";
+import Map from "./Map";
+import WorldButton from "./WorldButton";
+
+export default function Test() {
+  return (
+    <>
+      <h1>Test</h1>
+      <WorldButton />
+      <BiconomyButton />
+      <IPFSUploader />
+      <Map />
+
+      <FarmerCard />
+    </>
+  );
+}
