+++ packages/nextjs/pages/_app.tsx
@@ -8,7 +8,7 @@ import { Toaster } from "react-hot-toast";
 import { useDarkMode } from "usehooks-ts";
 import { WagmiConfig } from "wagmi";
 import { Footer } from "~~/components/Footer";
-import { Header } from "~~/components/Header";
+import Header from "~~/components/Header";
 import { BlockieAvatar } from "~~/components/scaffold-eth";
 import { useNativeCurrencyPrice } from "~~/hooks/scaffold-eth";
 import { useGlobalState } from "~~/services/store/store";
@@ -58,7 +58,7 @@ const ScaffoldEthApp = ({ Component, pageProps }: AppProps) => {
           theme={isDarkTheme ? darkTheme() : lightTheme()}
         >
           <div className="flex flex-col min-h-screen">
-            {/* <Header /> */}
+            <Header />
             <main className="relative flex flex-col flex-1">
               <Component {...pageProps} />
             </main>
